requirements.txt
minimalmodbus>=2.1.1
pymodbus>=3.7.2
pymongo>=4.10.1
pyserial>=3.5
python-dotenv>=1.0.1
Quart>=0.19.6
quart-cors>=0.7.0
tzlocal>=5.2